JWT 验证及代码优化

+++ app/model/model.go
@@ -1,8 +1,8 @@
 package model
 
 type Model struct {
-	ID        uint `gorm:"primary_key" json:"id"`
-	CreatedAt int  `json:"created_at"`
-	UpdatedAt int  `json:"updated_at"`
-	DeletedAt int  `gorm:"index"`
+	ID        uint64 `gorm:"primary_key" json:"id"`
+	CreatedAt int    `json:"created_at"`
+	UpdatedAt int    `json:"updated_at"`
+	DeletedAt int    `gorm:"index"`
 }
